pdf-operations/extract_pdf_tables.py
import argparse
import pandas as pd
from pdfminer.high_level import extract_pages
from pdfminer.layout import LTTextContainer, LTChar, LAParams

def extract_tables_from_pdf(input_path, output_path, output_format="csv"):
    """
    Extracts tables from a PDF and saves them in the specified format.
    This is a simplified table extraction and may not work for all PDF layouts.

    Args:
        input_path (str): The path to the input PDF file.
        output_path (str): The path to save the extracted table(s).
        output_format (str, optional): The output format ("csv" or "json"). Defaults to "csv".
    """
    try:
        laparams = LAParams(line_margin=0.2, char_margin=2.0, boxes_flow=0.5)
        all_tables = []

        with open(input_path, 'rb') as f:
            for page_layout in extract_pages(f, laparams=laparams):
                # A simple heuristic: find clusters of text that look like a table
                # This is a placeholder for a more robust table detection algorithm
                # For a real-world scenario, a library like `camelot-py` or `tabula-py` would be better.
                # Since we are constrained to the listed dependencies, we'll do a basic extraction.

                # This basic implementation will extract text lines and their positions.
                # A more complex logic is needed to assemble these into tables.
                # For now, we will just extract all text and assume it's one big table per page.

                page_text = []
                for element in page_layout:
                    if isinstance(element, LTTextContainer):
                        page_text.append(element.get_text())

                # This is a very naive approach, assuming each line is a row.
                if page_text:
                    df = pd.DataFrame([row.strip().split() for row in page_text if row.strip()])
                    all_tables.append(df)

        if all_tables:
            # Concatenate all found tables
            result_df = pd.concat(all_tables, ignore_index=True)

            if output_format.lower() == "csv":
                result_df.to_csv(output_path, index=False)
                print(f"Successfully extracted tables to '{output_path}' in CSV format.")
            elif output_format.lower() == "json":
                result_df.to_json(output_path, orient="records", indent=4)
                print(f"Successfully extracted tables to '{output_path}' in JSON format.")
            else:
                print(f"Error: Unsupported output format '{output_format}'. Please use 'csv' or 'json'.")
        else:
            print("No tables found in the PDF.")

    except FileNotFoundError:
        print(f"Error: Input file not found at '{input_path}'")
    except Exception as e:
        print(f"An unexpected error occurred: {e}")

def main():
    """
    Main function to parse command-line arguments and run table extraction.
    """
    parser = argparse.ArgumentParser(
        description="Extract tables from a PDF file. Note: This is a best-effort extraction."
    )
    parser.add_argument(
        "input_file",
        metavar="INPUT_PDF",
        type=str,
        help="The path to the PDF file from which to extract tables.",
    )
    parser.add_argument(
        "-o",
        "--output",
        dest="output_file",
        type=str,
        required=True,
        help="The path to save the extracted table data.",
    )
    parser.add_argument(
        "-f",
        "--format",
        dest="output_format",
        type=str,
        default="csv",
        choices=["csv", "json"],
        help="The output format for the table data (csv or json).",
    )

    args = parser.parse_args()

    extract_tables_from_pdf(args.input_file, args.output_file, args.output_format)

if __name__ == "__main__":
    main()