pdf-operations/requirements.txt
pypdf
pdfminer.six
reportlab
Pillow
pdf2image